Move VMSA definitions into the SNP crate
