+++ cpuarch/src/lib.rs
@@ -1,3 +1,4 @@
 #![no_std]
 
 pub mod snp_cpuid;
+pub mod vmsa;
